Move authOptions and token refresh functions out of route.ts file and into their own files
