+++ src/app/hooks/useSession.ts
@@ -1,6 +1,6 @@
 import { getServerSession } from "next-auth"
 
-import { authOptions } from "../api/auth/[...nextauth]/route"
+import { authOptions } from "@/utilities/OAuth/authOptions"
 
 export const useSession = async () => {
   const session = await getServerSession(authOptions)
