+++ src/app/page.tsx
@@ -4,8 +4,7 @@ import { getServerSession } from "next-auth/next"
 import { LoginButton, LogoutButton } from "@/components/buttons"
 import { SongCard } from "@/components/songCard"
 import { songs } from "@/data/songs"
-
-import { authOptions } from "./api/auth/[...nextauth]/route"
+import { authOptions } from "@/utilities/OAuth/authOptions"
 
 export default async function Home() {
   const session = await getServerSession(authOptions)
