+++ src/lib/posts.ts
@@ -1,7 +1,7 @@
 // 'use server'
 
 // import { getServerSession } from "next-auth";
-// import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+// import { authOptions } from "@/utilities/OAuth/authOptions"
 // import prisma from "@/lib/prisma";
 // import { revalidatePath } from "next/cache";
