+++ src/lib/prisma/suggested.ts
@@ -3,7 +3,7 @@
 import { Album, Artist, Image, Song, Suggested } from "@prisma/client"
 import { getServerSession } from "next-auth"
 
-import { authOptions } from "@/app/api/auth/[...nextauth]/route"
+import { authOptions } from "@/utilities/OAuth/authOptions"
 import { addToQueue, getSong } from "@/utilities/spotifyAPI"
 
 import prisma from "../prisma"
